pyproject.toml
[tool.poetry]
name = "onlinetradingplatform"
version = "0.1.0"
description = ""
authors = ["ZhenyaKasatkina <knopisha@list.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
django-relatives = "^1.3.0"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
coverage = "^7.6.1"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.4.0"
flake8 = "^7.1.1"
black = "^24.8.0"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
